# docs/api-contract.yaml
openapi: 3.0.3
info:
  title: Jingxin-Agent API
  version: 1.0.0
  description: |
    Jingxin-Agent 智能对话系统 API 规范

    **基础信息**:
    - Base URL: `https://api.jingxin.example.com`
    - 认证方式: Bearer Token
    - 响应格式: JSON (除 SSE 流式接口外)

    **版本控制**:
    - v1: 当前稳定版本
servers:
  - url: https://api.jingxin.example.com
    description: Production
  - url: https://staging-api.jingxin.example.com
    description: Staging
  - url: http://localhost:3001
    description: Local Development
security:
  - BearerAuth: []
...
paths:
...
  /files/{file_id}:
    get:
      tags: [Files]
      summary: 下载文件
      parameters:
        - name: file_id
          in: path
          required: true
          schema:
            type: string
        - name: mode
          in: query
          schema:
            type: string
            enum: [direct, presigned]
            default: direct
          description: |
            - direct: 直接下载文件
            - presigned: 返回签名 URL
      responses:
        '200':
          description: 成功 (direct 模式返回文件内容，presigned 模式返回 JSON)
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          download_url:
                            type: string
                          expires_at:
                            type: string
                            format: date-time
                          expires_in:
                            type: integer
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  # ==================== User APIs ====================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
    ApiResponse:
      type: object
      required: [code, message, trace_id]
      properties:
        code:
          type: integer
          description: 业务状态码
          example: 10000
        message:
          type: string
          description: 响应消息
          example: "Success"
        data:
          type: object
          nullable: true
          description: 业务数据
        trace_id:
          type: string
          description: 请求追踪 ID
          example: "req_abc123def456"
        timestamp:
          type: integer
          description: 响应时间戳 (Unix 毫秒)
          example: 1707811200000
...
  responses:
...
    Forbidden:
      description: 无权限
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/ApiResponse'
              - type: object
                properties:
                  code:
                    type: integer
                    example: 31001
                  message:
                    type: string
                    example: "Access denied"
    NotFound:
      description: 资源不存在
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/ApiResponse'
              - type: object
                properties:
                  code:
                    type: integer
                    example: 40001
                  message:
                    type: string
                    example: "Resource not found"
...